changed /events to /meetings
