+++ package.json
@@ -14,10 +14,12 @@
     "markdown-it": false
   },
   "dependencies": {
+    "@vueuse/head": "^1.3.1",
     "body-parser": "^1.19.0",
     "cross-env": "^7.0.3",
     "express": "^4.17.1",
     "express-session": "^1.17.3",
+    "i": "^0.3.7",
     "markdown-it": "^13.0.2",
     "multer": "^1.4.5-lts.1",
     "passport": "^0.6.0",
