+++ src/components/SocialMedia.vue
@@ -1,7 +1,5 @@
 <template>
   <div class="footer">
-    <a class="social" href="https://www.facebook.com/groups/hacksu"><img src="../assets/facebook-white.svg"></a>
-    <a class="social" href="https://twitter.com/_hacksu"><img src="../assets/twitter-white.svg"></a>
     <a class="social" href="https://www.instagram.com/hacksu/"><img src="../assets/instagram-white.svg"></a>
     <a class="social" href="https://github.com/hacksu"><img src="../assets/github-padded-white.svg"></a>
     <a v-if="discord" class="social" v-bind:href="discord"><img src="../assets/discord-square-white.svg"></a>
